+++ classifai-core/src/main/java/ai/classifai/action/LabelListImport.java
@@ -16,7 +16,6 @@
 package ai.classifai.action;
 
 import lombok.Getter;
-import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.IOUtils;
 
@@ -37,8 +36,14 @@ public class LabelListImport
     @Getter
     private List<String> validLabelList = null;
 
-    public LabelListImport(@NonNull File labelFile)
+    public LabelListImport(File labelFile)
     {
+        if(!labelFile.exists())
+        {
+            validLabelList = new ArrayList<>();
+            return;
+        }
+
         try
         {
             String labels = IOUtils.toString(new FileReader(labelFile));